+++ examples/hybrid_rrf/model.rb
@@ -1,11 +1,5 @@
 # frozen_string_literal: true
 
-begin
-  require "neighbor"
-rescue LoadError
-  abort "Example requires neighbor. Run with `BUNDLE_GEMFILE=examples/Gemfile bundle exec ...`."
-end
-
 require "active_record"
 require_relative "../../lib/parade_db"
 
@@ -14,8 +8,6 @@ class MockItem < ActiveRecord::Base
 
   self.table_name = "mock_items"
   self.primary_key = "id"
-
-  has_neighbors :embedding
 end
 
 class MockItemIndex < ParadeDB::Index